pyproject.toml
[tool.poetry]
name = "tabim"
version = "0.1.0"
description = "Guitar Pro to ASCII Tab converter"
authors = ["Tamir Bahar"]
license = "MIT"

packages = [
    { include = "tabim", from = "src" },
]

[tool.poetry.scripts]
tabim = 'tabim.main:main'


[tool.poetry.dependencies]
python = "^3.7"
pyguitarpro = "^0.8"
attrs = "^21.2.0"
more-itertools = "^8.8.0"
typer = "^0.4.0"
approvaltests = "^2.0.0"
pytest-approvaltests = "^0.2.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.8b0"
isort = "^5.9.3"
mypy = "^0.910"
autoflake = "^1.4"
structlog = "^21.1.0"
rich = "^10.9.0"
ipython = "^7.27.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.isort]
profile = "black"
src_paths = ["src/tabim"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --approvaltests-use-reporter='PythonNative' --approvaltests-subdirectory='approved_files'"
testpaths = [
    "tests",
]
